+++ R/keep.R
@@ -20,6 +20,9 @@
 #'
 #' @seealso [keep_at()]/[discard_at()] to keep/discard elements by name.
 #' @param ... Additional arguments passed on to `.p`.
+#' @param .progress Whether to show a progress bar. Use `TRUE` to turn on
+#'   a basic progress bar, use a string to give it a name, or see
+#'   [progress_bars] for more details.
 #' @export
 #' @examples
 #' rep(10, 10) |>
@@ -41,23 +44,23 @@
 #' # compact() discards elements that are NULL or that have length zero
 #' list(a = "a", b = NULL, c = integer(0), d = NA, e = list()) |>
 #'   compact()
-keep <- function(.x, .p, ...) {
-  where <- where_if(.x, .p, ...)
+keep <- function(.x, .p, ..., .progress = FALSE) {
+  where <- where_if(.x, .p, ..., .progress = .progress)
   .x[!is.na(where) & where]
 }
 
 #' @export
 #' @rdname keep
-discard <- function(.x, .p, ...) {
-  where <- where_if(.x, .p, ...)
+discard <- function(.x, .p, ..., .progress = FALSE) {
+  where <- where_if(.x, .p, ..., .progress = .progress)
   .x[is.na(where) | !where]
 }
 
 #' @export
 #' @rdname keep
-compact <- function(.x, .p = identity) {
+compact <- function(.x, .p = identity, .progress = FALSE) {
   .f <- as_mapper(.p)
-  discard(.x, function(x) is_empty(.f(x)))
+  discard(.x, function(x) is_empty(.f(x)), .progress = .progress)
 }