+++ R/pmap.R
@@ -98,28 +98,28 @@
 #' map2_dbl(df$x, df$y, min)
 #' pmap_dbl(df, min)
 pmap <- function(.l, .f, ..., .progress = FALSE) {
-  pmap_("list", .l, .f, ..., .progress = .progress)
+  pmap_(.type = "list", .l, .f, ..., .progress = .progress)
 }
 
 #' @export
 #' @rdname pmap
 pmap_lgl <- function(.l, .f, ..., .progress = FALSE) {
-  pmap_("logical", .l, .f, ..., .progress = .progress)
+  pmap_(.type = "logical", .l, .f, ..., .progress = .progress)
 }
 #' @export
 #' @rdname pmap
 pmap_int <- function(.l, .f, ..., .progress = FALSE) {
-  pmap_("integer", .l, .f, ..., .progress = .progress)
+  pmap_(.type = "integer", .l, .f, ..., .progress = .progress)
 }
 #' @export
 #' @rdname pmap
 pmap_dbl <- function(.l, .f, ..., .progress = FALSE) {
-  pmap_("double", .l, .f, ..., .progress = .progress)
+  pmap_(.type = "double", .l, .f, ..., .progress = .progress)
 }
 #' @export
 #' @rdname pmap
 pmap_chr <- function(.l, .f, ..., .progress = FALSE) {
-  pmap_("character", .l, .f, ..., .progress = .progress)
+  pmap_(.type = "character", .l, .f, ..., .progress = .progress)
 }
 
 pmap_ <- function(